# backend/docs/schemas/openapi.yaml
openapi: 3.0.0
info:
  version: 1.1.0
  title: HomiesHomes Real Estate API
  description: A RESTful API for HomiesHomes real estate listings platform. This API provides endpoints for managing users, properties, bookings, favorites, and agent requests.
  contact:
    name: Mandeep Pandya
  license:
    name: Apache 2.0
    url: "https://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://gammacairo-deltareward-9000.codio-box.uk"
    description: Codio API server
  - url: "https://gammacairo-deltareward-3000.codio-box.uk"
    description: Codio frontend server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT Authentication. A valid JWT token must be passed in the Authorization header.
        Example: `Authorization: Bearer {token}`

        You can obtain a token by using the /login endpoint with valid credentials.
        Tokens are valid for 1 hour after issuance.
  schemas:
    Error:
      type: object
      properties:
        status:
          type: integer
          example: 400
        error:
          type: string
          example: "Bad Request"
        message:
          type: string
          example: "Validation failed"
paths:
...
  /properties/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Property ID

    get:
      tags:
        - Properties
      summary: Get property by ID
      description: Returns a property by its ID.
      responses:
        "200":
          description: Property details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      id:
                        type: integer
                        example: 1
                      title:
                        type: string
                        example: "Modern Apartment Downtown"
                      description:
                        type: string
                        example: "Beautiful modern apartment in the heart of the city."
                      price:
                        type: number
                        example: 250000
                      location:
                        type: string
                        example: "Birmingham, UK"
                      bedrooms:
                        type: integer
                        example: 2
                      bathrooms:
                        type: integer
                        example: 1
                      property_type:
                        type: string
                        example: "Apartment"
                      image_url:
                        type: string
                        example: "prop1.jpg"
                      agent_id:
                        type: integer
                        example: 2
                      agent:
                        type: object
                        properties:
                          _id:
                            type: string
                            example: "60d0fe4f5311236168a109ca"
                          name:
                            type: string
                            example: "John Smith"
                          phone:
                            type: string
                            example: "07700900000"
                          avatar:
                            type: string
                            example: "avatar.jpg"
                  success:
                    type: boolean
                    example: true
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    put:
      tags:
        - Properties
      summary: Update property
      description: Updates a property listing (agent owner or admin role required).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  example: "Updated Apartment Title"
                description:
                  type: string
                  example: "Updated property description."
                price:
                  type: number
                  example: 275000
                location:
                  type: string
                  example: "Updated location"
                bedrooms:
                  type: integer
                  example: 3
                bathrooms:
                  type: integer
                  example: 2
                property_type:
                  type: string
                  example: "Condo"
                image_url:
                  type: string
                  example: "prop2.jpg"
      responses:
        "200":
          description: Property successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Property updated successfully"
                  property_id:
                    type: integer
                    example: 1
        "401":
          description: Unauthorized - Must be property owner or admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

    delete:
      tags:
        - Properties
      summary: Delete property
      description: Deletes a property listing (agent owner or admin role required).
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Property successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Property deleted successfully"
        "401":
          description: Unauthorized - Must be property owner or admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /bookings/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Booking ID

    get:
      tags:
        - Bookings
      summary: Get booking by ID
      description: Returns a booking by its ID. Users can only view their own bookings, agents can view bookings for their properties, and admins can view all bookings.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Booking details
          content:
            application/json:
              schema:
                type: object
                properties:
                  booking_id:
                    type: integer
                    example: 1
                  property_id:
                    type: integer
                    example: 1
                  user_id:
                    type: integer
                    example: 3
                  agent_id:
                    type: integer
                    example: 2
                  booking_status:
                    type: string
                    enum: [pending, confirmed, cancelled, attended]
                    example: "pending"
                  scheduled_date:
                    type: string
                    format: date
                    example: "2023-04-15"
                  scheduled_time:
                    type: string
                    example: "14:00"
                  notes:
                    type: string
                    example: "Interested in the kitchen space"
                  booking_created:
                    type: string
                    format: date-time
                    example: "2023-04-01T10:30:00Z"
                  property:
                    type: object
                    properties:
                      title:
                        type: string
                        example: "Modern Apartment Downtown"
                      location:
                        type: string
                        example: "Birmingham, UK"
                      image_url:
                        type: string
                        example: "prop1.jpg"
                  user:
                    type: object
                    properties:
                      username:
                        type: string
                        example: "john_doe"
                      email:
                        type: string
                        example: "john.doe@example.com"
                      name:
                        type: string
                        example: "John Doe"
                      phone:
                        type: string
                        example: "1234567890"
                  agent:
                    type: object
                    properties:
                      username:
                        type: string
                        example: "agent_smith"
                      email:
                        type: string
                        example: "agent.smith@example.com"
                      name:
                        type: string
                        example: "Agent Smith"
                      phone:
                        type: string
                        example: "0987654321"
                  _links:
                    type: object
                    properties:
                      self:
                        type: object
                        properties:
                          href:
                            type: string
                            example: "/bookings/1"
                          method:
                            type: string
                            example: "GET"
        "401":
          description: Unauthorized - Must be booking owner, property agent, or admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /bookings/{id}/status:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
        required: true
        description: Booking ID

    put:
      tags:
        - Bookings
      summary: Update booking status
      description: Updates a booking's status. Users can cancel their own bookings, agents can update statuses for their properties, and admins can update all booking statuses.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [pending, confirmed, cancelled, attended]
                  example: "confirmed"
      responses:
        "200":
          description: Booking status successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Booking status updated to confirmed"
                  booking_id:
                    type: integer
                    example: 1
                  status:
                    type: string
                    example: "confirmed"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...